+++ Pipfile
@@ -18,7 +18,9 @@ itsdangerous = "*"
 Jinja2 = "*"
 jmespath = "*"
 MarkupSafe = "*"
+pandas = "*"
 python-dateutil = "*"
+plotly = "*"
 s3transfer = "*"
 six = "*"
 urllib3 = "*"
