+++ maps_app/utils/sampledata.py
@@ -1,302 +1,90 @@
-items = [
-    {
-        "latlng": "41.86745834459586,(-78.58319325808752,)",
-        "lat": 41.86745834459586,
-        "lng": -78.58319325808752,
-        "name": "York Voting Center",
-        "city": "York",
+import random
+
+random.seed = 10
+
+
+def generate_locations(num, lat=40.4406, lng=-79.9959, current_count=0, locations=[]):
+    if current_count == num:
+        return locations
+    # PA_bounds = {"north": 42.3, "south": 39.7, "west": -80.5, "east": -74.7}
+
+    # Adjust the latitude and longitude slightly for each subsequent location
+    new_lat = random.randint(40, 41) + random.random()
+    new_lng = (random.randrange(77, 80) + random.random()).__neg__()
+
+    # Mock address and name for the location
+    address_num = random.randint(100, 999)
+    city_names = [
+        "Pittsburgh",
+        "Philadelphia",
+        "Harrisburg",
+        "Lancaster",
+        "Allentown",
+        "Erie",
+        "Scranton",
+        "Bethlehem",
+        "State College",
+        "York",
+        "Reading",
+        "Altoona",
+        "Wilkes-Barre",
+        "Easton",
+        "Lebanon",
+        "Hazleton",
+        "Chambersburg",
+        "Pottsville",
+        "Carlisle",
+        "Hanover",
+        "Williamsport",
+        "Sharon",
+        "Hermitage",
+        "Greensburg",
+        "New Castle",
+        "Johnstown",
+        "McKeesport",
+        "Norristown",
+        "Chester",
+        "Bethel Park",
+        "Monroeville",
+        "Plum",
+        "Doylestown",
+        "Meadville",
+        "Indiana",
+        "St. Marys",
+    ]
+    street_names = [
+        "Liberty",
+        "Broad",
+        "Penn",
+        "Main",
+        "Hamilton",
+        "College",
+        "Lakefront",
+        "King",
+        "Capitol",
+        "Market",
+    ]
+    street_types = ["St", "Ave", "Dr", "Rd", "Blvd"]
+    city = random.choice(city_names)
+    street_name = random.choice(street_names)
+    zip_code = random.randint(15001, 19640)
+    address = f"{address_num} {street_name} {random.choice(street_types)}, {city}, PA {zip_code}"
+    name = f"{city} Voting Center"
+
+    location = {
+        "latlng": f"{new_lat},{new_lng}",
+        "lat": new_lat,
+        "lng": new_lng,
+        "name": name,
+        "city": city,
         "state": "PA",
-        "address": "194 Main Ave, York, PA 16296",
-        "zip": 16296,
-    },
-    {
-        "latlng": "40.955522161135065,(-78.4618437611921,)",
-        "lat": 40.955522161135065,
-        "lng": -78.4618437611921,
-        "name": "Chester Voting Center",
-        "city": "Chester",
-        "state": "PA",
-        "address": "923 Liberty Rd, Chester, PA 19561",
-        "zip": 19561,
-    },
-    {
-        "latlng": "40.955621966756944,(-77.23031550275704,)",
-        "lat": 40.955621966756944,
-        "lng": -77.23031550275704,
-        "name": "Pottsville Voting Center",
-        "city": "Pottsville",
-        "state": "PA",
-        "address": "732 Liberty Ave, Pottsville, PA 15780",
-        "zip": 15780,
-    },
-    {
-        "latlng": "40.7903763002592,(-79.03959123041253,)",
-        "lat": 40.7903763002592,
-        "lng": -79.03959123041253,
-        "name": "Altoona Voting Center",
-        "city": "Altoona",
-        "state": "PA",
-        "address": "112 College Ave, Altoona, PA 16787",
-        "zip": 16787,
-    },
-    {
-        "latlng": "41.53645482429568,(-77.73894928748686,)",
-        "lat": 41.53645482429568,
-        "lng": -77.73894928748686,
-        "name": "Greensburg Voting Center",
-        "city": "Greensburg",
-        "state": "PA",
-        "address": "307 Hamilton Dr, Greensburg, PA 18467",
-        "zip": 18467,
-    },
-    {
-        "latlng": "41.153867990617094,(-77.18484339425183,)",
-        "lat": 41.153867990617094,
-        "lng": -77.18484339425183,
-        "name": "Wilkes-Barre Voting Center",
-        "city": "Wilkes-Barre",
-        "state": "PA",
-        "address": "722 College Rd, Wilkes-Barre, PA 19530",
-        "zip": 19530,
-    },
-    {
-        "latlng": "41.379735572012564,(-77.03122395370428,)",
-        "lat": 41.379735572012564,
-        "lng": -77.03122395370428,
-        "name": "Lebanon Voting Center",
-        "city": "Lebanon",
-        "state": "PA",
-        "address": "631 Broad Blvd, Lebanon, PA 15404",
-        "zip": 15404,
-    },
-    {
-        "latlng": "40.50247020286659,(-77.68031930734307,)",
-        "lat": 40.50247020286659,
-        "lng": -77.68031930734307,
-        "name": "Sharon Voting Center",
-        "city": "Sharon",
-        "state": "PA",
-        "address": "953 Penn St, Sharon, PA 16344",
-        "zip": 16344,
-    },
-    {
-        "latlng": "41.157935446018215,(-79.12308698130066,)",
-        "lat": 41.157935446018215,
-        "lng": -79.12308698130066,
-        "name": "Philadelphia Voting Center",
-        "city": "Philadelphia",
-        "state": "PA",
-        "address": "461 Main Dr, Philadelphia, PA 17365",
-        "zip": 17365,
-    },
-    {
-        "latlng": "40.04985133385288,(-79.35173577157204,)",
-        "lat": 40.04985133385288,
-        "lng": -79.35173577157204,
-        "name": "Sharon Voting Center",
-        "city": "Sharon",
-        "state": "PA",
-        "address": "576 Penn Dr, Sharon, PA 18503",
-        "zip": 18503,
-    },
-    {
-        "latlng": "40.052820414216214,(-79.06514193539353,)",
-        "lat": 40.052820414216214,
-        "lng": -79.06514193539353,
-        "name": "Bethel Park Voting Center",
-        "city": "Bethel Park",
-        "state": "PA",
-        "address": "878 Market Dr, Bethel Park, PA 15068",
-        "zip": 15068,
-    },
-    {
-        "latlng": "41.068471247554925,(-77.68410968926948,)",
-        "lat": 41.068471247554925,
-        "lng": -77.68410968926948,
-        "name": "Greensburg Voting Center",
-        "city": "Greensburg",
-        "state": "PA",
-        "address": "167 Penn Dr, Greensburg, PA 19341",
-        "zip": 19341,
-    },
-    {
-        "latlng": "40.1078907528757,(-78.00256380405712,)",
-        "lat": 40.1078907528757,
-        "lng": -78.00256380405712,
-        "name": "St. Marys Voting Center",
-        "city": "St. Marys",
-        "state": "PA",
-        "address": "324 Lakefront Dr, St. Marys, PA 16029",
-        "zip": 16029,
-    },
-    {
-        "latlng": "41.83812540086425,(-78.76493900239936,)",
-        "lat": 41.83812540086425,
-        "lng": -78.76493900239936,
-        "name": "Lebanon Voting Center",
-        "city": "Lebanon",
-        "state": "PA",
-        "address": "161 Broad Dr, Lebanon, PA 19343",
-        "zip": 19343,
-    },
-    {
-        "latlng": "41.1576320966243,(-78.8442306790118,)",
-        "lat": 41.1576320966243,
-        "lng": -78.8442306790118,
-        "name": "Bethlehem Voting Center",
-        "city": "Bethlehem",
-        "state": "PA",
-        "address": "628 College Ave, Bethlehem, PA 15536",
-        "zip": 15536,
-    },
-    {
-        "latlng": "41.60584100521868,(-78.3401084554076,)",
-        "lat": 41.60584100521868,
-        "lng": -78.3401084554076,
-        "name": "Scranton Voting Center",
-        "city": "Scranton",
-        "state": "PA",
-        "address": "208 Liberty Rd, Scranton, PA 16192",
-        "zip": 16192,
-    },
-    {
-        "latlng": "40.69388214336153,(-77.17681726359145,)",
-        "lat": 40.69388214336153,
-        "lng": -77.17681726359145,
-        "name": "Meadville Voting Center",
-        "city": "Meadville",
-        "state": "PA",
-        "address": "555 Penn Blvd, Meadville, PA 19477",
-        "zip": 19477,
-    },
-    {
-        "latlng": "41.04867707930188,(-78.81822220970601,)",
-        "lat": 41.04867707930188,
-        "lng": -78.81822220970601,
-        "name": "Johnstown Voting Center",
-        "city": "Johnstown",
-        "state": "PA",
-        "address": "769 Market Ave, Johnstown, PA 18352",
-        "zip": 18352,
-    },
-    {
-        "latlng": "41.85427424700172,(-77.16551089527175,)",
-        "lat": 41.85427424700172,
-        "lng": -77.16551089527175,
-        "name": "St. Marys Voting Center",
-        "city": "St. Marys",
-        "state": "PA",
-        "address": "558 Main Rd, St. Marys, PA 15071",
-        "zip": 15071,
-    },
-    {
-        "latlng": "41.15022156427101,(-79.74495838030094,)",
-        "lat": 41.15022156427101,
-        "lng": -79.74495838030094,
-        "name": "St. Marys Voting Center",
-        "city": "St. Marys",
-        "state": "PA",
-        "address": "279 Market Ave, St. Marys, PA 17896",
-        "zip": 17896,
-    },
-    {
-        "latlng": "40.94481897014843,(-79.60373148705345,)",
-        "lat": 40.94481897014843,
-        "lng": -79.60373148705345,
-        "name": "Norristown Voting Center",
-        "city": "Norristown",
-        "state": "PA",
-        "address": "139 Capitol Dr, Norristown, PA 16119",
-        "zip": 16119,
-    },
-    {
-        "latlng": "41.29689376351061,(-77.55067749619117,)",
-        "lat": 41.29689376351061,
-        "lng": -77.55067749619117,
-        "name": "Norristown Voting Center",
-        "city": "Norristown",
-        "state": "PA",
-        "address": "576 Broad Rd, Norristown, PA 19177",
-        "zip": 19177,
-    },
-    {
-        "latlng": "41.43067855130117,(-78.85269083699046,)",
-        "lat": 41.43067855130117,
-        "lng": -78.85269083699046,
-        "name": "Scranton Voting Center",
-        "city": "Scranton",
-        "state": "PA",
-        "address": "728 King Blvd, Scranton, PA 19411",
-        "zip": 19411,
-    },
-    {
-        "latlng": "41.11076855961876,(-78.97283027850297,)",
-        "lat": 41.11076855961876,
-        "lng": -78.97283027850297,
-        "name": "Harrisburg Voting Center",
-        "city": "Harrisburg",
-        "state": "PA",
-        "address": "798 Market St, Harrisburg, PA 15599",
-        "zip": 15599,
-    },
-    {
-        "latlng": "41.13831032057414,(-79.84744047868476,)",
-        "lat": 41.13831032057414,
-        "lng": -79.84744047868476,
-        "name": "Hanover Voting Center",
-        "city": "Hanover",
-        "state": "PA",
-        "address": "979 Penn Blvd, Hanover, PA 17109",
-        "zip": 17109,
-    },
-    {
-        "latlng": "40.2304581406016,(-78.3453723284815,)",
-        "lat": 40.2304581406016,
-        "lng": -78.3453723284815,
-        "name": "Indiana Voting Center",
-        "city": "Indiana",
-        "state": "PA",
-        "address": "763 Main Blvd, Indiana, PA 17381",
-        "zip": 17381,
-    },
-    {
-        "latlng": "40.26255949098571,(-77.8703380786689,)",
-        "lat": 40.26255949098571,
-        "lng": -77.8703380786689,
-        "name": "Chester Voting Center",
-        "city": "Chester",
-        "state": "PA",
-        "address": "945 Penn Rd, Chester, PA 18945",
-        "zip": 18945,
-    },
-    {
-        "latlng": "41.42627398426417,(-77.06787816407893,)",
-        "lat": 41.42627398426417,
-        "lng": -77.06787816407893,
-        "name": "Easton Voting Center",
-        "city": "Easton",
-        "state": "PA",
-        "address": "890 Broad St, Easton, PA 16569",
-        "zip": 16569,
-    },
-    {
-        "latlng": "40.349588306149,(-78.02929603327235,)",
-        "lat": 40.349588306149,
-        "lng": -78.02929603327235,
-        "name": "Pittsburgh Voting Center",
-        "city": "Pittsburgh",
-        "state": "PA",
-        "address": "379 Penn Ave, Pittsburgh, PA 16684",
-        "zip": 16684,
-    },
-    {
-        "latlng": "40.9592854326863,(-78.24676887951404,)",
-        "lat": 40.9592854326863,
-        "lng": -78.24676887951404,
-        "name": "Erie Voting Center",
-        "city": "Erie",
-        "state": "PA",
-        "address": "125 King St, Erie, PA 17491",
-        "zip": 17491,
-    },
-]
+        "address": address,
+        "zip": zip_code,
+    }
+
+    locations.append(location)
+
+    return generate_locations(num, new_lat, new_lng, current_count + 1, locations)
+
+
+voting_locations = generate_locations(15)
